Separate groups from scim endpoints
