+++ src/frontegg-mock/src/models.rs
@@ -7,12 +7,14 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+pub mod group;
 pub mod scim;
 pub mod sso;
 pub mod token;
 pub mod user;
 pub mod utils;
 
+pub use group::*;
 pub use scim::*;
 pub use sso::*;
 pub use token::*;
